+++ backend/routes/Saniru/gameRoutes.js
@@ -7,5 +7,6 @@ router.get('/', gameController.getAllGames);
 router.post('/createGame', gameController.createGame);
 router.delete('/deleteGame/:gameId', gameController.deleteGame);
 router.put('/updateGame/:gameId', gameController.updateGame);
+router.get('/:gameId', gameController.getGameById);
 
 module.exports = router;
