+++ frontend/package.json
@@ -14,6 +14,8 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.6.8",
     "firebase": "^10.11.0",
+    "jspdf": "^2.5.1",
+    "jspdf-autotable": "^3.8.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.22.3",
