--- frontend/src/pages/Game library/game.js
@@ -1,155 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import Header from '../../Components/common/Header/header';
-import Footer from '../../Components/common/Footer/footer';
-import { Box } from '@mui/material';
-import { styled } from '@mui/material/styles';
-import ButtonBase from '@mui/material/ButtonBase';
-import Typography from '@mui/material/Typography';
-import { NavLink } from 'react-router-dom';
-import axios from 'axios'; // Import Axios
-
-const ImageButton = styled(ButtonBase)(({ theme }) => ({
-  position: 'relative',
-  height: 200,
-  width: 200,
-  margin: '30px',
-  border: '1px solid #FB41FF',
-  borderBottom: 'none',
-  variant: 'outlined',
-  [theme.breakpoints.down('sm')]: {
-    width: '100% !important',
-    height: 100,
-  },
-  '&:hover, &.Mui-focusVisible': {
-    zIndex: 1,
-    '& .MuiImageBackdrop-root': {
-      opacity: 0.55,
-    },
-    '& .MuiImageMarked-root': {
-      opacity: 0,
-    },
-    '& .MuiTypography-root': {
-      border: '3px solid #FFFF',
-      variant: 'outlined',
-      color: '#FFFFFF',
-    },
-  },
-}));
-
-const ImageSrc = styled('span')({
-  position: 'absolute',
-  left: 0,
-  right: 0,
-  top: 0,
-  bottom: 0,
-  backgroundSize: 'cover',
-  backgroundPosition: 'center 40%',
-});
-
-const Image = styled('span')(({ theme }) => ({
-  position: 'absolute',
-  left: 0,
-  right: 0,
-  top: 0,
-  bottom: 0,
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  color: theme.palette.common.white,
-}));
-
-const ImageBackdrop = styled('span')(({ theme }) => ({
-  position: 'absolute',
-  left: 0,
-  right: 0,
-  top: 0,
-  bottom: 0,
-  backgroundColor: theme.palette.common.black,
-  opacity: 0.1,
-  transition: theme.transitions.create('opacity'),
-}));
-
-const ImageMarked = styled('span')(({ theme }) => ({
-  height: 3,
-  width: 18,
-  backgroundColor: theme.palette.common.white,
-  position: 'absolute',
-  bottom: -2,
-  left: 'calc(50% - 9px)',
-  transition: theme.transitions.create('opacity'),
-}));
-
-const Game = () => {
-  // State to hold the games data
-  const [games, setGames] = useState([]);
-
-  useEffect(() => {
-    // Fetch games data from backend using Axios
-    axios.get('http://localhost:5000/api/games')
-      .then((response) => {
-        setGames(response.data);
-      })
-      .catch((error) => {
-        console.error('Error fetching games:', error);
-      });
-  }, []);
-
-  return (
-    <Box>
-      <Header />
-      <Box
-        sx={{
-          backgroundImage: `url('https://images.saymedia-content.com/.image/t_share/MTkzNzg4MTIxMjM2NjQ1MzE1/aesthetic-website-backgrounds.gif')`,
-          backgroundSize: 'cover',
-          backgroundPosition: 'center',
-          minHeight: '100vh',
-          display: 'flex',
-          flexDirection: 'column',
-          alignItems: 'center',
-          justifyContent: 'flex-start',
-        }}
-      >
-        <h1>Game Library</h1>
-        <Box
-          sx={{
-            display: 'flex',
-            flexDirection: 'row',
-            flexWrap: 'wrap',
-            justifyContent: 'center',
-          }}
-        >
-          {games.map((game) => (
-            <NavLink key={game.gameId} to={`/gamedetails/${game.gameId}`} style={{ textDecoration: 'none' }}>
-              <ImageButton focusRipple>
-                <ImageSrc style={{ backgroundImage: `url(${game.image})` }} />
-                <ImageBackdrop className="MuiImageBackdrop-root" />
-                <Image>
-                  <Typography
-                    component="span"
-                    variant="subtitle1"
-                    sx={{
-                      position: 'absolute',
-                      bottom: '10px',
-                      fontSize: '12px',
-                      color: 'rgba(107, 107, 107, 0.5)',
-                      p: 4,
-                      pt: 2,
-                      pb: (theme) => `calc(${theme.spacing(1)} + 6px)`,
-                      textDecoration: 'none',
-                    }}
-                  >
-                    {game.title}
-                    <ImageMarked className="MuiImageMarked-root" />
-                  </Typography>
-                </Image>
-              </ImageButton>
-            </NavLink>
-          ))}
-        </Box>
-      </Box>
-      <Footer />
-    </Box>
-  );
-};
-
-export default Game;
